pyproject.toml
[tool.poetry]
name = "ultravox"
version = "0.1.0"
description = ""
authors = ["Fixie.ai Team <founders@fixie.ai>"]
packages = [{ include = "ultravox/" }]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
torch = ">=2.4"
transformers = {version = ">=4.43.1", extras = ["torch"]}
bitsandbytes = "~0.42.0"
peft = "~0.11.1"
simple-parsing = "~0.1.5"
librosa = "~0.10.2.post1"
requests = "~2.31.0"
datasets = "~2.19.1"
mosaicml-streaming = "~0.7.6"
nltk = "~3.8.1"
truecase = "~0.0.14"
sentencepiece = "~0.2.0"
protobuf = "~4.25.3"
dataclasses-json = "~0.6.7"
openai = "~1.33.0"
jiwer = "~3.0.4"
tensorboardx = "~2.6.2.2"
wandb = "~0.17.1"
sacrebleu = "^2.4.2"
tenacity = "^9.0.0"
evals = {git = "https://github.com/fixie-ai/evals"}

[tool.poetry.group.dev.dependencies]
black = "~24.4.2"
isort = "~5.13.2"
mypy = "~1.10.0"
autoflake = "~2.3.1"
pytest = "~8.2.2"
fsspec = "~2024.3.1"
gcsfs = "~2024.3.1"
sounddevice = "~0.4.7"
mosaicml-cli = "~0.6.31"
gradio-client = ">=0.16.1"
gradio = ">=4.29.0"
gpustat = "~1.1.1"
types-requests = "~2.26.0"
types-pyyaml = "^6.0.12.20240724"


[tool.poetry.group.apisvc.dependencies]
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
structlog = "^24.4.0"
python-dotenv = "^1.0.1"
websockets = "<13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.mason.api"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.isort]
profile = "black"
single_line_exclusions = ["typing", "collections.abc", "typing_extensions"]
skip = [
  "venv",
  ".venv",
  "third_party",
]

[tool.black]
extend-exclude = '''
/(
  | third_party
  | venv
)/
'''
